refactor(blocks): move format-color out of internal/
