+++ packages/blocks/src/TokenDetail.tsx
@@ -1,6 +1,6 @@
 import type { ReactElement } from 'react';
 import { useColorFormat } from '#/contexts.ts';
-import { formatColor } from '#/internal/format-color.ts';
+import { formatColor } from '#/format-color.ts';
 import { formatValue } from '#/internal/use-project.ts';
 import { AliasChain } from '#/token-detail/AliasChain.tsx';
 import { AliasedBy } from '#/token-detail/AliasedBy.tsx';
