+++ packages/blocks/src/internal/channel-globals.ts
@@ -1,6 +1,6 @@
 import { useSyncExternalStore } from 'react';
 import { addons } from 'storybook/preview-api';
-import { COLOR_FORMATS, type ColorFormat } from '#/internal/format-color.ts';
+import { COLOR_FORMATS, type ColorFormat } from '#/format-color.ts';
 
 /**
  * Shared subscription to Storybook's globals channel, lifted out of React
