+++ packages/blocks/src/token-detail/AxisVariance.tsx
@@ -1,7 +1,7 @@
 import type { ReactElement } from 'react';
 import { useMemo } from 'react';
 import { useColorFormat } from '#/contexts.ts';
-import { type ColorFormat, formatColor } from '#/internal/format-color.ts';
+import { type ColorFormat, formatColor } from '#/format-color.ts';
 import { formatValue } from '#/internal/use-project.ts';
 import { styles } from '#/token-detail/styles.ts';
 import {
